+++ packages/mobile/src/components/input/index.ts
@@ -1,2 +1,2 @@
-export * from './types';
-export * from './text-input';
+export * from './text-input/types';
+export * from './text-input/text-input';
